AFNetworking実装途中

+++ hack418/hack418/Model/MyMeetLogItem.swift
@@ -11,9 +11,9 @@ import Foundation
 struct MyMeetLogItem {
     var name: String!
     var image_url: String!
-    var position: Double!
+    var position: Float!
     
-    init(name: String, image_url: String, position: Double) {
+    init(name: String, image_url: String, position: Float) {
         self.name = name
         self.image_url = image_url
         self.position = position
